Add prettier (#99)
This PR adds prettier to the EFNYC codebase and uses it to update the formatting across all relevant files. It also adds yarn prettier:check to our yarn test command, which now acts as its sole test for now 😞. Once we have CircleCI enabled, we will be sure to run yarn test in order to establish prettier as a requirement for all incoming PRs to the codebase. Fixes #84.