Add Typescript and a barebones Jest test suite (#101)
This PR adds a barebones Jest configuration to allow for basic test suites. It's modeled after JustFixNYC/justfix-website@9171e57.

This PR also adds typescript and ts configurations as a regular dependency. We needed this now so we could make sure our Jest configuration is compatible with .tsx components, which we will inevitably be working with after converting some components to typescript.

Fixes #85 and fixes #93.