Add scaffolding for productionalized serverless functions. (#116)
This adds support for productionalized serverless functions, starting with an example `hello-world` function.

The function is written in TypeScript, bundled via [Rollup][] using the same Babel configuration we use for Jest, and deployed via [Netlify Functions][], which means it's accessible at `/.netlify/functions/hello-world`.  It can be passed `?kaboom=on` to make it throw an exception with source-mapped line numbers in the traceback.

The function can be accessed via development at the same location, but one needs to run `yarn netlify:develop` to start up their dev server, and the dev server is accessed at `http://localhost:8001/` rather than on port 8000.  Any changes to the function will be instantly accessible on the server (no need to restart the server manually).

In production, any exceptions raised by the function are reported to Rollbar, assuming the `ROLLBAR_POST_SERVER_ITEM_ACCESS_TOKEN` is set in the Gatsby configuration.

It also adds a test suite.

I feel we can use this in production.  I don't think it couples us _too_ much to Netlify, since the way serverless functions are done is the exact same as AWS Lambda (they even use AWS Lambda under the hood), and all the other services that provide serverless are generally compatible with it.  It would be a bit of a hassle to move off Netlify if we start using their functions, sure, but it wouldn't be a deal-breaker by any means.

[Rollup]: https://rollupjs.org/
[Netlify Functions]: https://docs.netlify.com/functions/overview/