Refactor our "Send to Phone" feature using a serverless function (#127)
This PR refactors our "Send to Phone" feature on the court results page to make use of our new Netlify serverless functions infrastructure. In doing so, this PR adds a few more env vars for configuration and also removes the need for an APIClient file, as our front-end now only communicates with our own defined functions.